ShapeProgramSE4.001
First upload of shape program for SE4, applied inheritance and polymorphism. 
Created an abstract Shape class with the following properties of colour, x position and y position which other shapes can inherit.
So far just developed working code that implements a range of shapes such as circle, rectangle and square which all inherit the Shape class and contain an overidden draw method. 
Version 001