ShapeProgramSE4.004
Version 004
Created Parser class, began some development on the concept of it and what needs to be required. So far the parser class takes input from the user, removes excess spaces, converts to lowercase and splits up input into a range of arrays for further processing. More functionality needs to be applied however I have tested and debugged the existing functionality and it has logic and performs as it should. 
Tidied up code in some areas and added more xml documentation for clear commenting of code.