ShapeProgramSE4.006
ShapeProgramSE4.006
Added a CommandFactory Class to follow factory design pattern. CommandFactory allows dependency of all the commands to come from this one place. 
I have also altered the Parser Class so each time a command is inputted from the user the program will create an object of the specific command using the CommandFactory and then split up the command into parameters to process them. I have also made a Class for each command, so far this includes DrawTo and MoveTo which are both draw commands. I have also created a class called DrawCommands which extends the Command class and all draw commands must have a canvas. Each command will contain methods to ParseParameters and ProcessParameters in order for the specified command to be drawn on the Canvas. Each command will also contain a Set and Execute method to allow Parameters to be passed around the program effectively.
I have added a Keywords class as extra functionality to allow the Graphical Programming Application to act like more of a programming environment where the user can recognise keywords throughout the program. This feature allows keywords to change colour when the user is typing in either 2 boxes on the application. For example; when the user types 'moveto' or 'drawto' the word will highlight green.

I have also created a GPLException Class where exceptions will be stored and referenced in other classes. This just needs implementing for the next version.
I have also set up a UnitTest application which will be developed further in the upcoming versions. This will allow me to test each class and function effectively.
Actions for next versions: is to implement mor requirements and complete unfinished classes.
Think about including another design pattern.