ShapeProgramSE4.007
ShapeProgramSE4.007
I have created a few more classes that extend the drawcommand class such as 'DrawRectangle', 'DrawCircle' etc. These classes allow shapes to be drawn to the canvas. I have also added short command classes such as clear, pencolor, reset which are still extended by the drawcommand as they control actions on the canvas. I added in a variety of set commands in drawcommands to compensate for short commands that require less variables passed in. This functionality is tested and works. 
Also added in a few exceptions to handle the program exceptions that should be raised if the user for example doesnt type in a command correctly - these exceptions are created from GPLException class. 
I have implemented save and load program features however this saves and loads the actual canvas of the program and not the text. This could be an extra feature but unsure yet, this will definitely be corrected in the next version so a file of the program is saved instead.
Also tweaked functionality of command box. User can now type as much lines of code as they want in here and then press run to execute the program. Still need to add in functionality for when user types run code is executed (without the need to press 'Run' button.
Added in var (variable) class and introduced declaration and setting of variables to work in program.  At the moment not a lot is been done with the Var class apart from setting and getting variables - most of the work happens in the parser class where arrays are created to store the name and value of variables and checking to see if variable exists before one is created/set also happens in parser class. - Need to play around to see if this can be moved elsewhere as code is looking a bit unreadable.

Next need to todo list:
-Add in methods declaration and setting and look at the way I have applied this to see if it is sound and logical. 
-Add in if statements.
-Add in loops
-Make sure syntax is checked before code is ran
-Start unit testing.
-Try to add any additional functionality