ShapeProgramSE4.008
ShapeProgramSE4.008
Since the last upload I have fixed a few problems.
I have corrected the functionality of the save and load buttons from the toolbar to allow the user to be able to save and load the code as text a text file. I moved the canvas save and load functionality as an additional feature to my program, which allows the user to save and load the canvas to the program. 
Created a CodeChecker class that is responsible for validating the code as the user types a line in the command box. 
Created method class and need to finish off implementation
Actions for next upload.
- Complete method implementation
- Tidy up code
- Finish off exceptions
- Start unit testing 
- Add in loops - study iterator pattern
- Attempt if statement