ShapeProgramSE4.013
ShapeProgramSE4.013
Added unit tests to this version - broke down files.
In this upload I have introduced the iterator pattern when methods have been declared. This allows the key idea in this pattern is to take the responsibility for access and traversal out of the list object and put it into an iterator object. I now need to work on applying this to variables/loops. First I will begin with getting while loops working and using the iterator pattern.

Actions for next upload.
- See if methods with parameters can be implemented
- Add in loops - study iterator pattern/implement while loops
- Add more unit testing
- Finalise exceptions
- Tidy up code